FFTObj class moved to own .cpp and .h file
