+++ CMakeLists.txt
@@ -111,6 +111,7 @@ add_library(genesis13 STATIC
         src/Util/Sorting.cpp
         src/Util/StringProcessing.cpp
         src/Util/SemaFile.cpp
+        src/Util/FFTObj.cpp
         src/Lattice/AlterLattice.cpp
         src/Lattice/Lattice.cpp
         src/Lattice/LatticeElements.cpp
