+++ include/Diagnostic.h
@@ -19,6 +19,7 @@
 
 #ifdef FFTW
 #include <fftw3.h>
+#include "FFTObj.h"
 #endif
 
 
@@ -74,36 +75,6 @@ public:
     virtual void getValues(Beam *, std::map<std::string,std::vector<double> > &, int) =0;
 };
 
-
-
-// For field diagnostics: helper class for management of FFT-related resources
-#ifdef FFTW
-class FFTObj {
-public:
-	FFTObj(int);
-	~FFTObj();
-
-	// The fftw_plan is a pointer, delete copy constructor and copy
-	// assignment operator.
-	// This avoids dangling pointers in object copies (fftw_plan contains
-	// the data locations that were used to prepare the plan).
-	FFTObj(const FFTObj&) = delete;
-	FFTObj& operator= (const FFTObj&) = delete;
-
-	std::complex<double> *in_ {nullptr};
-	std::complex<double> *out_ {nullptr};
-
-	// according to FFTW documentation, fftw_plan is an "opaque pointer type" (https://www.fftw.org/fftw3_doc/Using-Plans.html , 19.01.2023)
-	fftw_plan p_;
-
-private:
-	int rank_;
-	int ngrid_;
-};
-// map holding FFT-related resources for the values of ngrid
-typedef std::map<int,FFTObj *> map_fftobj;
-#endif
-
 class DiagFieldBase{
 protected:
     std::map<std::string, OutputInfo>  tags;  // holds a map with tags and units
@@ -118,9 +89,6 @@ public:
     DiagFieldBase() = default;
     virtual std::map<std::string,OutputInfo> getTags(FilterDiagnostics &filter) =0;
     virtual void getValues(Field *, std::map<std::string,std::vector<double> > &, int) =0;
-//#ifdef FFTW
-//    virtual void destroyFFTPlan()=0;
-//#endif
 };
